Create a Generic DiagLogger Interface for Collecting Radio Logs
This controller defines an interface that can be used by
on-device logging mechanisms that capture logs for offline
analysis. The semantics support basic start/stop, pulling
of logs, and resetting the logging for subsequent tests. Because
the mechanisms for this are typically proprietary, vendors must
provide their own instantiations of the DiagLogger in order to
collect logs.

Bug: 27299974
Change-Id: Ie2ff6ce7e3fcdbdb78f3bea9f6cd87629d6adc73