PBAP SL4A testing.
Test scripts to support automated testing of PBAP functionality for the
car kit.  This test currently expects a paired phone/car kit, erases
contacts out of the phone, uploads a new contact file to the phone, and
then connects the phone with the car kit and verifies that the contacts
were transfered correctly.

bug: 28794836
Change-Id: I28b971d3b404e1ccf2d7fd2e825c1dffc16115ad