Wait for Android device to boot before taking a bug report
In instances where a test failure is the result of a runtime
restart or a reboot, taking a bug report will fail because it
is attempted before the device's ADB interface is available.
To avoid this, we should wait for the device to reboot/restart
prior to taking a bug report.

Bug: 29216085
Change-Id: Ic2ad35b95aa6b84e43b67c48fcf265bbb1b409f0