Add fuzz testing for HFP
Does quick dial/hangup tests to stress the various components of
telephony. The motivation is that different delays involve different
level of interaction between telephony and bluetooth which can lead to
race condition if not given enough time to stabilize. The test case
tests the effect of very short, medium and large delays on how HFP
calling reacts.

Bug: b/29101689
Change-Id: I4eea0f3532daa88099457842808a4ee32f256a9a