Clean up Android Devices if Failure to Create Device 2+
When creating AndroidDevice objects, if any device fails
after it has launched an additional subprocess, an exception
causing an exit() condition fails to clean up the subprocess
and causes the parent process to hang on exit(). This CL
performs graceful cleanup of all AndroidDevices including the
current insofar as the devices have threads/subprocesses
associated with them.

Bug: 29636143
Change-Id: Iffa0b19935b17ce70cfbdffbf74ea4e9da091270