[DO NOT MERGE] Add the config path to the config dictionary.
With autotest the config file is put into a temporary directory.
Since autotest is cloning files in a directory it makes sense to
have addition files be moved over to the drone.

These files will be sitting next to the config file, so if we
can easily get the folder of the config file then we can use
these other files.

TEST=Ran it
BUG=b/29836695

Change-Id: Ie29c0e2874f901af765b483ba16bbc9361ee74d3