Test scanning for LE devices after BT reset
This test is to verify that LE devices are found
successfully after the Bluetooth stack is
reset. This is repeated multiple times in order
to verify that LE devices are not lost in scanning
when the stack is brought back up.

Added common functions for concurrent advertising
in bt_test_utils.py. Modified existing
ConcurrentBleAdvertisingTest to use it.

Bug: b/29960087
Change-Id: Ia26bed2d6a21cff7f8658ab59a7df5d5c29a409b