Patch Wi-Fi Autojoin and Scanner BSSID test to make them work on new test setup.
To enable the most amount of tests in new setup without invoking
dramatic refactor of the code, we introduced a new param to singal
which type of setup we are running on. The param can be set in
config files.

We also forked the Autojoin tests for the new setup.

Other bug fixes:
Fix execution order in BSSID test.
Fix a missing param issue in WifiManagerTest.

Bug=30092848
Bug=30131989
Bug=30242975

Change-Id: I6cff15cc380b7d21d3d18115a405561da1a0d14c