wifi_test_utils use startTethering
Use the newly exposed ConnectivityManager.startTethering (and
stopTethering) calls to control the wifi hotspot in tests.

Note: further improvements will be made in later CLs to improve
logging and notifications about where the test is failing if the
hotspot does not come up fully.

The cherrypick for this CL was not clean, this is a new CL and has
been modified slightly from what was checked in to mr1
(uses asserts.assert). I also attempted to test against a master
build, but settings and other components were repeatedly crashing
and I was unable to even fully lauch the test reliably.

BUG: 30112957
Change-Id: I3b7d0114af7c278937d6bb07abca24adb12e7ac5