Remove most tests from preflight checking.
There are two problems with the current preflight test, which
are leading me to temporarily remove most of its provisioning
checks.

1) We now have a combination of a device and carrier that each
support a feature, but where that feature isn't supported on
that device, which has exposed the (known) flaw in the current
logic: its assumption is no longer valid.

2) The logic, when re-written, needs to be genericized to allow
scalability and to add provisioning checks for new features.

Temporarily removing it, and at the same time, ensuring that all
failures will now result in an abort-all behavior.

Bug: 30228034
Change-Id: I5c32a62bfce52023b30a79f1749096a08c409948