[DO NOT MERGE] WifiSoftApTest: new tests for wifi tethering
This CL adds a new set of tests for softap mode.  These tests verify
each wifi component needed for tethering.
1 - calls to check if wifi tethering is supported
2 - starting softap (wifi only - no tethering/connectivity)
3 - full wifi tethering started (sim not required)

The tests check that wifi is operating as expected, not that
connectivity is working.  The second DUT is used to check for the
soft AP in scan results and is not currently connecting to the softap.

BUG: 30907130
TEST: ran locally on shamu and bullhead

Change-Id: Ib0eb1488170a8e3c174080822a568d888862f733