Fix some instances of stress_test_number KeyError
Add a method get_stress_test_number in TelephonyBaseTest that returns
the integer value of the stress_test_number parameter, or None if it was
not set and logs a warning. Modifies subclasses of TelephonyBaseTest
to use the get_stress_test_number method instead of keying directly into
the user_params dict.

Change-Id: I523ceb9c9a6dd8b576e5e7f224c7c0aac213b3a7