Fixed up ssh code
Ssh code will now do a safty check on the output to make sure
it is coming from the remote machine. Commands will append a
unique hash to be echoed before commands start to run. Run
now will only report errors when this hash is not found
in the output.

Run will now check using the MULTILINE flag for regex. This
prevents errors where the output was not on a single line
as expected.

Run will now raise an error if none of the expected
errors occured, previously just returned None.

Formatter will now only export variables when variables are
given. This prevents export from being called and printing
the current enviroment at the start of the standard output.

Bug: 31687844
Change-Id: I970d7fb9b54a2ac74e1297b988a3d32808b8a597