Refactored background job to more clear location.
Background job now sits in job.py to be less redundent

job.py moved to acts/libs/proc so it can be used as a
standard tool in all parts of acts.

Refactored class names such as CmdError to be less
reduendent with module name.

Background job is now in a subfolder called proc for better
organization.

Refactored all code that had dependencies.

Bug: 31689075
Test: None
Change-Id: Id02f67dcd7a0af9b8c4b59bc8de3e5ada65cd933