Update preferredNetworkType based on current setting
-Regardless of the preferred network type, there is an
early exit condition that caused us not to change the
preferred network type when the current network type
matched the requested network type. The change ensures
that the preferred network type is updated if necessary
even if the device is currently camped on the desired
type of network.

-Updated tel_test_utils with a newer version of YAPF.

Bug: 31594185
Change-Id: I251d1b7a8d61f125883b118c519a4e24a28e404f
(cherry picked from commit 72f9bffcdacbc670726cbe966331b2bf67465a65)