Default all attenuators to max value
It's possible that mobility tests that incorporate more
wireless feeds than are used. To avoid problems with
unanticipated signals in the tests, we now perform a
blanket set to max value on all attenuators. We must
assume that they are capable of attenuating the signals
they provide out of a usable/visible range.

Bug: 31556268
Change-Id: I87ecc89103855bb798b2122f53fd61097fdcae28