Refactor wait_for_ringing_event() to block on wait_for_event()
This function was using a hybrid of blocking and polling due to
some historical dropped events. Unfortunately, the code
introduces a race wherein the desired event comes in between
the time that the intermediate blocking occurs and the
check-for-error-condition code is executed, which was not
cognizant of the overall wait time. The simplest/best solution
is to simply rely on the blocking waiting for the event. While
this has the side effect of potentially blocking too long if the
test framework is buggy, we cannot design test code around buggy
infrastructure, as in this case the complexity introduced a
new subtle bug.

Bug: 31757791
Change-Id: I8f18f08ec5e90222a0e1c54406750217bce9830a