Make Airplane Mode Toggle Wait for All Radios to Change State
The toggle_airplane_mode() function previously only waited for
a connectivity state change, and then it would check that both
wifi and bluetooth radios had changed state. Unfortunately, those
radios sometimes require more time to turn on/off when entering
or leaving airplane mode than is taken by the CONNECTIVITY_CHANGE
broadcast. Thus, we need to wait for connectivity change and then
each of the peripheral radios in turn up to the maximum amount of
time.

Bug: 31473466
Test: none
Change-Id: I397d04f12cbd368ba229cb72306229580d76cb57