Added default timout to sl4a_client.open and fixed mac hang
Mac was hanging with sl4a_client.open because when the timeout
was None a large timeout was used instead. On mac this would cause
and integer overflow. Open now uses None with checks inside the
timeout loop instead to avoid this problem.

Open now has a default timeout of 1 minute to prevent hanging
forever.

Put in _counter variable again as it seems to have been deleted causing
acts to crash.

Bug: 32320104
Test: Ran unit tests.
Change-Id: I5b1b7dafac7c9e8a1150d2ce2b102589772c496c