Extend ACTS to run against remote testbeds
With this patch, AndroidDevice controllers can be passed an
optional config dictionary "ssh_config" that tells ACTS that
the Android device is attached to a remote host rather than
the local host.  For example:

...
"AndroidDevice": [
    {"serial": "xxxxxxxxxxxxxxxxxxxxx",
     "label": "bullhead",
     "ssh_config": {
        "user": "root",
        "host": "some.host.name.example.com"}
    }
],
...

This patch adds the logic to make sure adb versions are in sync,
that port forwarding works across devices, and that fastboot
runs against the remote host.

Bug: 32436082
Test: tests/sample/SampleTest passes against a remote testbed
Change-Id: If3ff364fd177338b3778b3aff372a5817664f775