Refactor AccessPoint to track running instances
AccessPoint will now track the instances of hostapd it is currently
running as a list (self._aps).  This allows us to generallize more
easily to heterogenous hardware situations.

While here, explicitly start and stop the associated DHCP server
with the subnets we expect to need rather than a hardcoded list.

Bug: 32571319
Test: An example test using configurable APs correctly brings up
      an AccessPoint with this code.

Change-Id: I3476dd11068b3e00e5855291558d561c0b45033f