WifiServiceApiTest: test remove network
Create two new tests to check if saved networks can be deleted with wifi
enabled/disabled.

Note: deleting a saved network when wifi is disabled is not implemented,
to work around this, the test was altered to pass with the current
behavior.  When b/32809235 is fixed, this test will fail and the
instructions in the comments in this test should be implemented.

Bug: 33683632
Test: ran new test locally
Change-Id: I0402406af95c850c0cb3ca8f55c1a66247707be4