Possible fix for sl4a client side port issue.
The tcp port forward would not work occasionally because the port was
already being listend to. This was most likely due to a race condition
with how we check if a port is free. This tries to get arround the
issue by first waiting for a small fraction of time after getting the
port to allow the OS time on the processor to reset the port as open,
and if all else fails trying to find another free port.

Bug: 33390806
Test: Ran long iteration of SimpleTest.

Change-Id: Iee7843ee7cf9c510d959dda73e7ef900825e95c0