Removed tdls_models and energy_info_models as necessary config items
Added the tdls_models and energy_info_models into the utils lib
since this does not change and does not need to be located in
in the configs. This was only for released devices so for
unreleased items the new optional config items have been add:
additional_energy_info_models and additional_tdls_models.

Bug: 34624546
Change-Id: I38e04f28ae687d859c3320611358728cd4c58104
Fixes: partially 34624546
Test: Ran test case with changes