Add special timeout for flashing and add workaround to fastboot
Flashing takes up to 3 mins so a special timeout needs to be set.
This is a workaround to get things working but the real change
will need to be done in
tools/test/connectivity/acts/framework/acts/controllers/fastboot.py
But this is a safer change in the short term.

Added a workaround to fastboot.py to take into account that
getvar outputs to stderr instead of stdout.  There is a bug
submitted for this b/34815412

Bug: 34773016
Change-Id: I7377bb3abf5099a8789d3bc46c173959b44e0d80
Fixes: 34773016
Test: Ran flashutil.py to flash devices.