Fix for logcat dying in the middle of a test
When logcat dies in the middle of a test, the test has no idea
when this happened and then at the end tries to clean up logcat
and then asserts since it is no longer running.  This changes it
so instead of just looking at the object to see if logcat is
available, the logcat pid is checked.  In addition, at the
beginning of each test logcat is confirmed to be running or
restarted if it is not.

Bug: 36556217
Change-Id: I7117993e258e110e7dd140e398c801d303fd7b12
Fixes: 36556217
Test: act.py -c /android/walleye.config -tc BleScanApiTest