Handle case where no android devices are being used in a test case
For the presubmit checks a mock test is used to verify the code,
unfortunately this does not have any android devices attached.
In this case, the previous check in tries to check the android
devices but since they don't exist, the mock test crashes.  Now,
the the code will check if there are any android devices before
trying to verify logcat is running or not.

Bug: 36790957
Change-Id: Ia7e23018cece175bbe9831e1f03a098485f7a75d
Fixes: 36790957
Test: ./acts/framework/tests/acts_base_class_test.py