Fixes for WifiScannerMultiScan & WifiManagerTest
Added the following:

WifiScannerMultiScan:
1. Time gap between two consecutive scan results should be less than 20%
of interval specified by scan settings. One of the recent run of this
script saw the difference slightly above 20%. Ningyuan suggested that as
long as this difference does not exceed 40% it should be fine. Changing
this to 25% as mentioned in the bug

WifiManagerTest:
1. Test throws exception due to self.current_test_name being None type

Combining into 1 CL as these are tiny fixes

These changes are cherry picked from aosp

Bug: 36593444, 36572881
Test: Verified manually
Change-Id: Id0a753de02dd521379a9cedee156bd4c65cd5154