[Tests]Wifi-Telephony coex and stress tests.
Added tests to ensure wifi and cellular are functional after doing
a sequence of events that alter wifi and cellular states. All tests
will be run on the Telephony rigs with sims and a wifi router.

The main idea of the tests is to connect to WiFi, setup voice calling,
then do toggle wifi and airplane and a combination of the two, and
ensure wifi is connectable and we can make cellular voice calls.

The stress tests will loop over each test, whth the loop count determined
from the impot supplied in the config file.

Bug: 36361712
Test: Tested locally

Change-Id: Ie1585937b49fe66e9b39e0d7f5ebdc875603ba94