Integrated relay controllers into ACTS
Added support for using relay controllers to interact with devices
used for testing. A device can be defined as a RelayDevice to work
nicely with the relay controllers without the test writer making low
level calls to switching on and off the relays. Different
RelayControllers can also be defined to allow different hardware
solutions ot be used to interact with the devices.

Bug: 36794752
Change-Id: I8fb095417e2d8da03dcfaf52b430d13f1d3aa74e
Fixes: 36794752
Test: acts/framework/tests/acts_relay_device_test.py
Test: ./RelayDeviceSampleTest -c <local config>