[AWARE] Self terminating sessions (TTL) with no term indication
When sessions self terminate they can either inform the host or not -
based on a configuration parameter (where off could be a power saver).
When they do not inform the host it cannot clean-up their state.
Therefore, if there is an attempt to reconfigure such a (no longer
existing) session it will result in a configuration failure (whereas
if state was updated we won't even try).

Update test script to expect a configuration failure in such a scenario.

Bug: 62054202
Test: DiscoveryTest:test_ttl* all pass
Change-Id: I45cbbf3fb50e423316a1fd14db73941f4b788e05