[AP_LIB]Fix for duplicate bssids
The code assumed that there is only 1 AP and if the bssid field is
not None, it would not generate new(virtual) bssid's, instead use
the existing bssid. For a multiple AP scenario new bssid should be
generated each time. The bssid check was added to support a fixed
mac but that is a highly unlikely test requirement.

Bug: 63400730
Test: Manual tests
Change-Id: I62b0cc664212b5a2d51f39b0b9b4cd803b232ef4