Add relative path resolution to relay_config.
Currently the RelayDevice value must be an absolute path.  However
with autotest and other test harnesses the paths will be in a
randomly generated folder.

This change uses the master config file location as the root
of the relative path generator.

In addition, there were several duplicated lines and also an
unreferenced variable that were removed.

Test: Manually, ran test cases:
  - Valid absolute path
  - Valid relative path
  - Invalid relative path
  - Invalid absolute path
Bug: 64981277

Change-Id: I29a772db63710d5000d03523d1887b1c43890ed3