[WifiPreFlightTest] Update reference network check to testcase
teardown_class does not get called if setup_class fails
As the setup_class fails when the test does not find the reference
networks when attn set to 0, the deletion of dynamic AP user params is
not called. As a result, the subsequent tests in nightly runs are
failing. Moving the reference network check to test case.

Bug: None
Change-Id: I945e2d848c365f88a4bf411583ed764c61bd86d1
Tests: Verified the changes