Add a user param to limit log output directory size.
The new user param "soft_output_size_limit" will specify an approximate
maximum size for the log output dir. It is labeled as approximate for two
reasons. First, it will only disable deivce bug reports, not all file
creation. Second, it only disables it after the limit has been reached,
so the actual size might be ~1 bug report larger than the specified max.

We implement this as only an approximate limit because we're only
looking to stop the output size from growing out of control (multiple
GB).

Bug: 68327895
Test: A failing version of SampleTest, along with multiple sample values
for the param.

Change-Id: Ic15e53d623b98c829067b17536ea352f27f4028a