Remove extraneous pip import from ACTS setup.
The pip import was initializing a bunch of submodules, including six,
which populated the module cache. If we didn't already explicitly
install another version, we ended up with 1.5.2 from /usr/share. When we
try to install 1.11.0, the setup script asks the six module what version
it is. Instead of the module we're installing, it receives the cached
six 1.5.2. This causes it to report that 1.5.2 gets installed, which
fails the acts installation, since the requirement was >=1.9.

We don't see this when six is manually installed, because the pip import
will load that version instead.

Bug: 69603278
Test: sudo pip3 uninstall six, sudo python3 setup.py develop(or install)
Change-Id: I91dfdfd6bed4508ec75fa20c07b3b45fe234ef08