WifiSoftApTest: Add tests for 2G & 5G band
The existing test uses the default band for testing tethering. Adding 2
versions of the test for explicitly testing AP bringup on 2G & 5G
bands.The 2G test is a duplicate of the default band test currently, but
this may change in the future where we're thinking of introducing a new
"Auto" band which would let the driver/firmware chose the best band.

Bug: 70269210
Test: Local test (`act.py -c softap.config -tb dut-name -tc
WifiSoftApTest`)
Change-Id: I1d1fac01af088f08b2460e53bef85ce4bf6e6922