WifiSoftApTest: Add hidden softap tests
Fixed couple of other nits,
a) The country code needs to be set on both the client & DUT for 5Ghz.
Also, use the WifiManager API to set country code. Setting it via halutil
may get overwritten by framework.
b) Ensure that we stop tethering when we exit this class.
c) Take bugreport for both client & DUT on failure.
d) Refactored the helper method to check if softAp AP is found or not in
scan results.

Bug: 63141761
Test: act.py -c softap.config -tb dut-name -tc WifiSoftApTest
Change-Id: Ic37e457ea6e4baafd462d903eccefa3efddbd5da