[AWARE] Test data-path with Responder from ANY configurations
Wi-Fi Aware data-path (NDP) setup requires an Initiator and a Responder.
Both Initiator and a Responder are set up with the MAC address of the
peer to which to connect to as well as a security configuration (Open,
PMK, Passphrase).

The original API (27) allowed a Responder to be configured to accept
a connection from anyone (e.g. using a null MAC address). This creates
ambiguous semantics when an NDP already exists. The second Responder
request could refer to the previously setup NDP or to a new one to be
created. We cannot tell the difference up-front.

While new API removes the "Accept request from ANYONE" - the tests are
configured to test with that ability. It provides coverage for API 27
(and existing code).

A test is adde to validate that for the current API level (most recent)
the Responder to ANY is not permitted.

Bug: 72175022
Test: integration tests
Change-Id: I78c2d8d059f6b12fd518501fabe7b2174b146458