wifi_test_utils: Handle scan failures
Changes in the CL:
a) Handle scan failures in the helper method
|start_wifi_connection_scan_and_check_for_network|. On failure, it'll
re-issue a new scan (until it reaches |max_tries|).
b) Increase the default value of |max_tries| to 3.
Note: This is a temporary fix, we should investigate & debug if we're
seeing 2 back to back scan failures.
c) Use the new utility method in
WifiManagerTest.connect_to_wifi_network_with_id.

Bug: 72659205
Test: `act.py -c wifi_manager.config -tb dut-name -tc WifiManagerTest`

Change-Id: Ief19bfbd9ecd2c9d99fcfa213c40ddd9b2138980