[Tests]SoftAP ACS tests
Added tests to test ACS on all 2.4 and 5ghz channels.
The test configures AP on a given channel. Then we start iPerf
traffic on that channel on the secondary device, during softAP
bring-up on the primary device. The primary device will bring-up
softAP in 2g/5g/auto bands.

Once the softAP is up, we uset he secondary device to connect to it
and get the softAP channel and verify that ACS did not pick the channel
that has active beacons on.

2.4GHz channels - 1,2,3,4,5,6,7,8,9,10,11
5GHz channels - 36,38,40,42,44,46,48,149,151,153,155,157,159,161,165

Also added a new flag, 'check_connectivity' to determine if we need to check
connectivity after connection. Needed to test softAP connection in rigs with
no SIMs. While here, also updated the frequency CHANNEL_MAP and commented out
all the DFS channels.

Bug: 72747921
Test: Tested on walleye. Few channels fail AP configuration that am looking
into: 38,42,46,151,155,159,165, but rest are good.
Change-Id: Ib7de898bb049b3e7e0197a07c30bb3d34d85fd1d