[AWARE] Configure discovery session with extra debugging information
Configure the discovery (publish/subscribe) sessions with the SSI set
to the name of the executing function. This does not impact discovery
but will help identify any Service Discovery indications with the method
to which they correspond. E.g. if we receive an extra Service Discovery
we can tell whether it is due to the current test or the previous
test (and just took a long time to arrive).

Bug: 33821639
Test: AwareDiscoveryWithRangingTest passes
Change-Id: I933a0148cc27c285e776be2823eda287937b4e89