Adding ability to use ADB devices as IPerfServers
This CL allows you to initiate an IPerfServer on an ADB device in much
the same way you can have an iperf server on a local or remote linux
machine. This improves readability and code reuse since we dont need
special handling (manual adb manipulation) in the test classes to run
iperf servers on android devices.

Test: Done
Bug: None

Change-Id: I1e7b9bc0363a09827f54e4a7b9fc5210773ab4d8
Signed-off-by: Omar El Ayach <oelayach@google.com>