[AWARE] Fix race condition between Responder and Initiator
Aware data-paths require that a Responder be configured before receiving
a request from the Initiator. The script "guarantees" this by making
the request on the Responder before making the request on the Initiator.
However, there are rare cases where either the host or the script or the
device introduced a delay - inverting the intialization sequence.

Add a 1 second delay to guarantee that the Responder is configured
before the Initiator.

Bug: 73896170
Test: DataPathStressTest passes even with much higher iteration count
Change-Id: Ia82ed963a7902a90e47ed64aeae4773cdcd0f75d