[AWARE] Relax requirements on concurrency test to allow duplicate matches
The Service Discovery callback (aka "match") is configured to only be
issued a single time per discovery. This "match once" is actually always
at best a guideline - several conditions can result in multiple matches
(e.g. device going out and into range). The concurency test triggers
another one of these conditions: changes in the underlying configuration
of the service discovery with ranging disable and re-enabled.

Relax the test to not fail on duplicate matches - as long as the
correct/expected ones are initially received.

Bug: 77925576
Test: AwareDiscoveryWithRangingTest:test_ndp_concurrency passes for partners
Change-Id: I391558c0d54b3ecf7ace096976505292b7bd79b6