[RTT] Update RTT ACTS to select 'best' APs for ranging
Current testing scanned for all APs of a particular type (e.g. RTT
supporting APs) and then ranged to all of them (possibly restricting
to the number of allowed parallel requests). This could result (in
an open environment) with picking ranging to distant APs over ranging
to nearer APs.

Modify tests to:
- Select the best (strongest RSSI) APs first.
- Limit the selected APs to have RSSI above a reasonable threshold.
- Limit the number of selected APs

Bug: 80319317
Test: all modified tests pass
Change-Id: Id08923066c415f3d3cf9a4a12951fbec117f0f1e