[WifiBaseTest]Configure same ssid on both bands and Un-mirror APs.
It is a not uncommon to have APs configured to have the same ssid
configured on both bands and the AP decide to pick one of them.
This can be used for new tests in teh future and also some network
selection tests.

With this support, it wouldnt make sense to always mirror
APs, then we will have just 1 ssid on nultiple bands and APs. Hence, added
the flexibility to configure eacj Ap separately, using two different configs
unlike the present setup to always mirroe APs. With this we can achieve
same ssid on each band and different ssid across APs. Some network selection
tests have this scenario.

Bug: 111366584
Bug: 72111271
Test: Locally. Verified wifimanger, autojoin and netowork selection tests.
Change-Id: I277cd332295629a7f9e4a065f2d839941f02d8f6