[RTT] Fix result reporting from test: extras must be a dict
Results are reported back from test as part of the 'extras' argument.
That argument must be a dict. Failure only starts (apparently) when
there is an attempt to upload the data to server - which only happens
when test is an official part of test suite (where 'official' == UUID
is present).

Simply converted result to a dictionary.

Bug: 112161056
Test: None
Change-Id: Ieacfea036099a6f11c65fb5d4fdaec5eb5ccb412