[Interop]Test case generation and testtracker upload.
Make changes so that the test uses ssid list to run through all
the possible ssids rahter than the APs. This is needed to make it
easy for test case generation which is per ssid. Changed the test
workflow, to create a base test method that is called for all the
generated testcases.

Bug: 112923784
Test: Local
Change-Id: I74c0d494a1b1f65343dcde79f0ed73707b80a738