Prepare tests for new DhcpServer implementation
Don't require hostname option not to be sent in DHCPOFFER: this is
inconsistent with DHCPACK behavior. The new DHCP server always sends the
option if a hostname was provided.

Don't require the broadcast flag to be set for replies to a renewing
query with the broadcast flag set. Such replies are done unicast, so
having the broadcast flag set is inconsistent, especially if the query
ends up relayed (the relay could broadcast it).

Test: tests pass
Bug: 109584964
Change-Id: I9ec8740b4301a144416c297b609e377fae3c2d6d