Fixes upload hooks and potential singleton module issues.
Importing modules with a relatvie path will register the module under
a different name, causing them to not share the same state.

This can effectively lead to static variables showing a different value
to different modules, depending on how it was imported. To prevent this,
we require all modules are imported with an absolute path.

Bug: None
Test: This upload passes upload hooks.

Change-Id: Ie6d954f500a5a91df50d59f1381e70bf5da1a0fb