Allow to choose number of UL/DL RBs on LTE
The 'pattern' parameter is now handled by the LteSimulation class to
decide the number of DL/UL RBs. The test name indicates the percentage
of RBs that have to be assigned to DL/UL and the simulation class finds
the number of DL/UL RBs that is closer to the requested percentages.

Bug: 112923220
Test: tested in the lab
Change-Id: Ief336c015843d73135eec05de3f60c55fe5ea5d2