Adds debug logging and diagnosis functionality to tnhelper
This CL adds debug logging and diagnosis functionality to the tnhelper
class. Commands that fail to return valid data are logged, along with
their result, and a diagnostic function is kicked off. The diagnostic
function attempts to ping the telnet server and reestablish
communication with it. Although the telnet commands may fail, the
diagnostic function, if successful, could get the telnet connection back
up for the next command.

Test: Done
Bug: None

Change-Id: I3b9047329e563cd79e84f22274cbe48a8385303f
Signed-off-by: Omar El Ayach <oelayach@google.com>