Starting the Wifi Power Tethering (Hotspot) Test Suite
This test suite measures the power consumption when Pixel is
configured in softAp mode (Hotspot). This class include idle tests
with/without client, traffic (downlink/uplink), WiFi Sharing (MCC/SCC
and DBS) and 2G/5G testing. This suite is a targeted subsystem measurement
in order to isolate the WiFi power contribution when Pixel is in Hotspot mode
since there is no cellular connection. This test suite is required in order
to catch and validate hotspot power before field deployment (like bug 117632438
and bug 111690416 which were found too late on the field)

Bug: None
Test: Tested in the lab
Change-Id: I17c410f009440dd30dccdc916bedf7c0ffbd6a54