Check throughput at the end of traffic tests
Clean up of the throughput check at the end of the tests. Moving the
check to the pass_fail_check method in the PowerTelTrafficTest class.
Also, modifying the pass_fail_check method in the base class to avoid an
exception when the threshold for power consumption in that test is not provided.

Bug: 112923220
Test: tested in the lab
Change-Id: I3826ae8ec018514c9f14a6579a38e8e71565d4eb