[AWARE] Add verification that network specifier info does not leak
The Aware network specifier may contain confidential information which
should not leak from service to framework. Add assertions to verify
that information is no leaked.

Note: this is part of a new capability to indicate to the framework not
to leak network specifiers to apps. The Aware network specifier was
created before this capability was added - so for now it obfuscates its
contents - i.e. not a real failure if leaks.

Bug: 122160111
Test: act.py DataPathTest
Test: act.py ThroughputTest
Change-Id: I391e46753897e04760e4132d9075ee5adcbc7833