Adding WiFi compatibility with the new scapy version
The scapy project released an new version for python3 which isn't
compatible with previous package. This fix updates the package name
and also the constants so that our code can work with any scapy
version. The latest official package is http://www.secdev.org/projects/scapy/
and scapy-python is deprecated

Bug: None
Test: Tested in the lab
Change-Id: I990b4b8f5766ab73145fd6e440ab7a7b26d6368f