WifiNetworkSuggestionTest: Modify suggestion remove behavior
To prevent apps from actively managing wifi connection via the new API,
we're not going to disconnect from the current network on suggestion
removal. We will just ensure that we don't connect to those in the
future. The current connection will linger on until it naturally
disconnects or if the device finds a better network to connect to (not
when the app tells us to).

Also, added a utility method to ensure there are no disconnects for a
specified duration.

Bug: 121268587
Test: act.py -c wifi_manager.config -tb dut-name -tc WifiNetworkSuggestionTest
Change-Id: If1edd9aa7cd0da5f25f16a87e2fff2ecd45ebc3e