Add capability to run OTA tests for Wifi
This CL includes a controller for OTA chambers which could include
turntables and stirrers. The controller is placed in the test_utils
folder and must be explicitly created in test classes that use it. The
current controller is a pseudo skeleton that supports most common use
cases in the OctoboxChamber. In the future this controller will be
extended to Bluetest and other chambers. The CL also includes a
WifiOTATest file which will define a number of test classes all of which
inherit from existing test classes but upgrade them to run in OTA
environments by setting turn table orientation for example.

Test: Done
Bug: None

Change-Id: Id08a386cd380d9896793e1ae6945a7d7d1bbfb8b
Signed-off-by: Omar El Ayach <oelayach@google.com>