Add single chain senstiivity tests
This CL adds the capability to do single chaing sensitivity tests by
attenuating the chain-not-in-test up to the maximum attenuation
supported by the attenuator. The CL also includes additional results
generation in a CSV file for better readability. To achieve the single
chain tests, the CL includes changes to the attenuator class to be more
permissive when setting attenuation levels above the maximum supported;
this is allowed via an optional flag passed to set_atten which defaults
to strict attenuation checking, i.e., current behavior.

Test: Done
Bug: None

Change-Id: I750a8abbb5867e22acfe215f82ba42b9bcd76207
Signed-off-by: Omar El Ayach <oelayach@google.com>