Updated find_executable to shutil.which.
The `which` command in bash skips executables in the path that the
current user does not have execute permissions for. find_executable
does not do this, which makes it more difficult to resolve why a
permission error may have been raised. shutil.which has the same
properties as `which`; this will guarantee protoc is taken from an
executable the user has permissions to run.

Bug: None
Test: Tested on a remote server.
Change-Id: I229ff267d09862c2dc4c95af5e4206e0a6ea6b7f