Fix test_max_concurrent_ble_scans_plus_one
Fix this test logic error. The verification should confirm that the
(max+1)th ScanCallback whether triggers onScanFailed or not, that is,
the (max+1)th LE scan should not start successfully.

These methods are unnecessary to be called: setting CallbackType, setting
ScanMode and building ScanFilter. Because this test not need to scan LE
advertisement and just verify the (max+1)th LE scan whether fail to start.

Bug: b/115307126
Test: Run SL4A ConcurrentBleScanningTest
Change-Id: I65773fa700486b7157ea85ebc5a35b7d7fcef0af