Test scripts for CellInfo Reform
Test script to verify RequestCellInfoUpdate()
/ GetAllCellInfo() API on Android Q
and regression check for GetAllCellInfo() on Android P

Test: Tested Locally creating a test campaign for Android Q and P and
running through all the test cases. A device with AOSP build was used for
the test with SL4A Q apk and P apk respectively.  The log info outputs
were used to view the results which also showed the system time stamps and
modem timestamps to ensure there are no false positives.

Bug: 124814885
Change-Id: Ia117da4b3672518ad5ed217e558ae5f57d929cad