Add plotting to wifi range over orientation
This CL compiled the test class results at the end of a run and plots
range for each channel over orientation in a single plot. It also
implements a mock OTA chamber object so that tests can be developed and
debugged on setups that do not have a chamber connected to them. The CL
also fixes a bug with output paths and test contexts.

Test: Done
Bug: None

Change-Id: I000ea92b8e15c8d25a3999508c3e03d3daa63605
Signed-off-by: Omar El Ayach <oelayach@google.com>