Make ACTS loggers point to their correct contexts upon initialization.
Previously, monolith, testclass, and testcase logs are all initialized to the current test context, as it is assumed to be the root context. This causes issues when one attempts to create a new logger during the middle of a test case, e.g. the new monolith logger would point to the testcase-level directory. This change fixes the issue.

Bug: 127438713
Test: act.py
Change-Id: Ie8ace552555742a3b7f0a6543ed12650098ab62e