Adding default downlink power for UMTS simulations
The method that allows to get uplink and downlink power values from the test name is being moved from LteSimulation to BaseSimulation so it can be used by UMTS sims too. At the same time, some changes in the variable names to allow for this method to be used for both sim types.

Bug: 131603324
Change-Id: Ibca4cc7099a2a8800a596573b7a4f0da2375656c