Made access_point controller work if the mac address had same last octet
There was an issue in the access point controller where if both
wlan interfaces had the same first byte both interfaces would
end up being the same because the last byte was always set to 0.
By changing it to start at 0 for one interface and 8 for the
this overlap doesn't happen.

Also fixed a minor bug in WlanScanTest where the teardown logic
was not being called.

Bug: None
Test: Ran the WlanScan test with an AP that had the same first byte
      in the last octet and made AP that had 8 ssids on one interface
      that went from hex 8-f.

Change-Id: I509ae2d9931cd4e7b69c6753faa9fbf09ff9bcf4